trying to merge manishs report
Using latex instead of writer because i like its look and we need to learn it if we want to write technical paper.
So i am learning it.

Changes to be committed:

deleted:    report.aux dont need to put it in repo automatically generated by latex
deleted:    report.log latex generated file
modified:   report.tex trying to merge your project with mine and trying to create project with standard report given to us.